feat(oas): merge example rules
