--- src/rulesets/oas/__tests__/oas3-valid-header-schema-example.ts
@@ -1,5 +1,0 @@
-import testRule from './templates/_schema-example';
-
-const ruleName = 'oas3-valid-header-schema-example';
-
-describe(ruleName, () => testRule(ruleName, 'headers'));
