--- src/rulesets/oas/__tests__/oas3-valid-oas-content-example.ts
@@ -1,5 +1,0 @@
-import testRule from './templates/_oas-example';
-
-const ruleName = 'oas3-valid-oas-content-example';
-
-describe(ruleName, () => testRule(ruleName, 'content'));
