+++ test-harness/scenarios/enabled-rules-amount.oas3.scenario
@@ -24,7 +24,7 @@ components:
 ====command====
 {bin} lint {document} --ruleset ./rulesets/parameter-description.oas3.yaml -v
 ====stdout====
-Found 54 rules (1 enabled)
+Found 48 rules (1 enabled)
 Linting {document}
 OpenAPI 3.x detected
