+++ test-harness/scenarios/examples.oas3.scenario
@@ -352,15 +352,15 @@ components:
 OpenAPI 3.x detected
 
 {document}
- 126:25  error  oas3-valid-schema-example            `example` property should have required property `name`
- 138:30  error  oas3-valid-schema-example            `example` property type should be boolean
- 142:30  error  oas3-valid-schema-example            `example` property should match format `date-time`
- 146:30  error  oas3-valid-schema-example            `example` property should match format `url`
- 169:20  error  oas3-valid-oas-parameter-example     `example` property should be equal to one of the allowed values: foo, bar
- 185:22  error  oas3-valid-parameter-schema-example  `example` property should be equal to one of the allowed values: foo, bar
- 197:22  error  oas3-valid-oas-parameter-example     `value` property type should be string
- 199:21  error  oas3-valid-oas-parameter-example     `value` property type should be string
- 262:25  error  oas3-valid-oas-content-example       `value` property should have required property `id`
- 282:25  error  oas3-valid-oas-content-example       `value` property should have required property `user`
+ 126:25  error  oas3-valid-schema-example  `example` property should have required property `name`
+ 138:30  error  oas3-valid-schema-example  `example` property type should be boolean
+ 142:30  error  oas3-valid-schema-example  `example` property should match format `date-time`
+ 146:30  error  oas3-valid-schema-example  `example` property should match format `url`
+ 169:20  error  oas3-valid-media-example   `example` property should be equal to one of the allowed values: foo, bar
+ 185:22  error  oas3-valid-schema-example  `example` property should be equal to one of the allowed values: foo, bar
+ 197:22  error  oas3-valid-media-example   `value` property type should be string
+ 199:21  error  oas3-valid-media-example   `value` property type should be string
+ 262:25  error  oas3-valid-media-example   `value` property should have required property `id`
+ 282:25  error  oas3-valid-media-example   `value` property should have required property `user`
 
 ✖ 10 problems (10 errors, 0 warnings, 0 infos, 0 hints)
